+++ lib/mito/coordinates/externals.h
@@ -10,6 +10,7 @@
 // externals
 
 // support
+#include "../journal.h"
 #include "../tensor.h"
